feat: support more delimiter for SSE events in GCP gemini (#1536)
**Description**
The SSE event spec supports 3 different delimiters (pair of CRLF / LF /
CR). This PR updates the stream processing translator for GCP VertexAI
to support all 3 delimiters.

[SSE docs] about delimiters
> **Note:** The docs define an event as `event = *( comment / field )
end-of-line`
`end-of-line` is defined as a single CRLF or LF or CR (not a pair)
But the `comment` and `field` definitions also end in `end-of-line`
implying that an event always ends in a pair of CRLR / CR / LF

[SSE docs]:
https://html.spec.whatwg.org/multipage/server-sent-events.html#parsing-an-event-stream

---------

Signed-off-by: Sukumar Gaonkar <sgaonkar4@bloomberg.net>
Co-authored-by: Ignasi Barrera <ignasi@tetrate.io>
Signed-off-by: yxia216 <yxia216@bloomberg.net>